backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
psycopg[binary]==3.2.3
PyJWT==2.10.1
passlib==1.7.4
